requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
channels==4.0.0
daphne==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1
# scikit-learn not yet available on Python 3.13; optional in code, fallback implemented
scikit-learn==1.3.2; python_version < "3.13"
dj-rest-auth==5.0.1
django-allauth==0.58.2
gunicorn==21.2.0
whitenoise==6.6.0
# psycopg2-binary is only needed for PostgreSQL; skip for local dev on SQLite/Windows
# psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
